+++ frontend/src/App.tsx
@@ -21,6 +21,7 @@ import ProfileContextProvider from "./contexts/ProfileContext";
 import { ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 import MatchProvider from "./contexts/MatchContext";
+import CollabProvider from "./contexts/CollabContext";
 import CollabSandbox from "./pages/CollabSandbox";
 import NoDirectAccessRoutes from "./components/NoDirectAccessRoutes";
 import EmailVerification from "./pages/EmailVerification";
@@ -30,55 +31,60 @@ function App() {
   return (
     <AuthProvider>
       <MatchProvider>
-        <Routes>
-          <Route path="/" element={<Layout />}>
-            <Route index element={<Landing />} />
-            <Route path="home" element={<ProtectedRoutes />}>
-              <Route index element={<Home />} />
-            </Route>
-            <Route path="questions">
-              <Route index element={<QuestionList />} />
-              <Route path=":questionId" element={<QuestionDetail />} />
-              <Route element={<ProtectedRoutes adminOnly />}>
-                <Route path="new" element={<NewQuestion />} />
-                <Route path=":questionId/edit" element={<QuestionEdit />} />
+        <CollabProvider>
+          <Routes>
+            <Route path="/" element={<Layout />}>
+              <Route index element={<Landing />} />
+              <Route path="home" element={<ProtectedRoutes />}>
+                <Route index element={<Home />} />
               </Route>
-            </Route>
-            <Route
-              path="profile/:userId"
-              element={
-                <ProfileContextProvider>
-                  <ProfilePage />
-                </ProfileContextProvider>
-              }
-            />
-            <Route path="profile/:userId/:qnHistoryId" element={<QuestionHistoryDetail />} />
-            <Route path="matching" element={<ProtectedRoutes />}>
-              <Route element={<NoDirectAccessRoutes />}>
-                <Route index element={<Matching />} />
-                <Route path="matched" element={<Matched />} />
-                <Route path="timeout" element={<Timeout />} />
+              <Route path="questions">
+                <Route index element={<QuestionList />} />
+                <Route path=":questionId" element={<QuestionDetail />} />
+                <Route element={<ProtectedRoutes adminOnly />}>
+                  <Route path="new" element={<NewQuestion />} />
+                  <Route path=":questionId/edit" element={<QuestionEdit />} />
+                </Route>
               </Route>
-            </Route>
-            <Route path="collaboration" element={<ProtectedRoutes />}>
-              <Route element={<NoDirectAccessRoutes />}>
-                <Route index element={<CollabSandbox />} />
+              <Route
+                path="profile/:userId"
+                element={
+                  <ProfileContextProvider>
+                    <ProfilePage />
+                  </ProfileContextProvider>
+                }
+              />
+              <Route
+                path="profile/:userId/:qnHistoryId"
+                element={<QuestionHistoryDetail />}
+              />
+              <Route path="matching" element={<ProtectedRoutes />}>
+                <Route element={<NoDirectAccessRoutes />}>
+                  <Route index element={<Matching />} />
+                  <Route path="matched" element={<Matched />} />
+                  <Route path="timeout" element={<Timeout />} />
+                </Route>
               </Route>
+              <Route path="collaboration" element={<ProtectedRoutes />}>
+                <Route element={<NoDirectAccessRoutes />}>
+                  <Route index element={<CollabSandbox />} />
+                </Route>
+              </Route>
+              <Route path="*" element={<PageNotFound />} />
             </Route>
-            <Route path="*" element={<PageNotFound />} />
-          </Route>
-          <Route path="auth">
-            <Route index element={<Navigate to="/auth/login" />} />
-            <Route path="signup" element={<SignUp />} />
-            <Route path="login" element={<LogIn />} />
-            <Route path="verify-email">
-              <Route index element={<EmailVerification />} />
-              <Route path=":userId" element={<EmailVerification />} />
+            <Route path="auth">
+              <Route index element={<Navigate to="/auth/login" />} />
+              <Route path="signup" element={<SignUp />} />
+              <Route path="login" element={<LogIn />} />
+              <Route path="verify-email">
+                <Route index element={<EmailVerification />} />
+                <Route path=":userId" element={<EmailVerification />} />
+              </Route>
+              <Route path="forget-password" element={<ForgetPassword />} />
+              <Route path="*" element={<Navigate to="/auth/login" />} />
             </Route>
-            <Route path="forget-password" element={<ForgetPassword />} />
-            <Route path="*" element={<Navigate to="/auth/login" />} />
-          </Route>
-        </Routes>
+          </Routes>
+        </CollabProvider>
       </MatchProvider>
       <ToastContainer position="bottom-right" />
     </AuthProvider>
