+++ frontend/src/components/CollabSessionControls/index.tsx
@@ -1,7 +1,7 @@
 import { Button, Stack } from "@mui/material";
 import Stopwatch from "../Stopwatch";
-import { useMatch } from "../../contexts/MatchContext";
-import { USE_MATCH_ERROR_MESSAGE } from "../../utils/constants";
+import { useCollab } from "../../contexts/CollabContext";
+import { USE_COLLAB_ERROR_MESSAGE } from "../../utils/constants";
 import { useEffect, useState } from "react";
 
 const CollabSessionControls: React.FC = () => {
@@ -16,11 +16,12 @@ const CollabSessionControls: React.FC = () => {
     return () => clearInterval(intervalId);
   }, [time]);
 
-  const match = useMatch();
-  if (!match) {
-    throw new Error(USE_MATCH_ERROR_MESSAGE);
+  const collab = useCollab();
+  if (!collab) {
+    throw new Error(USE_COLLAB_ERROR_MESSAGE);
   }
-  const { handleSubmitSessionClick, handleEndSessionClick } = match;
+
+  const { handleSubmitSessionClick, handleEndSessionClick } = collab;
 
   return (
     <Stack direction={"row"} alignItems={"center"} spacing={2}>
@@ -43,7 +44,9 @@ const CollabSessionControls: React.FC = () => {
         }}
         variant="outlined"
         color="error"
-        onClick={() =>{ handleEndSessionClick()}}
+        onClick={() => {
+          handleEndSessionClick();
+        }}
       >
         End Session
       </Button>
